+++ src/main/java/br/pucrio/les/esg_token_backend/services/value/IValueService.java
@@ -1,0 +1,5 @@
+package br.pucrio.les.esg_token_backend.services.value;
+
+public interface IValueService {
+
+}