+++ src/main/java/br/pucrio/les/esg_token_backend/services/value/ValueServiceImpl.java
@@ -1,0 +1,8 @@
+package br.pucrio.les.esg_token_backend.services.value;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ValueServiceImpl implements IValueService {
+
+}